+++ src/schemas/schema.d.ts
@@ -75,6 +75,10 @@ export interface APRenderRep {
          */
         strokeOpacity?: number;
         /**
+         * Only affects rect of hex maps. Used to fill hexes with a colour.
+         */
+        hexFill?: string;
+        /**
          * On `squares*` boards, blacks out the specified cells and disables clicking. For hex grids, the hex simply isn't drawn. Like with `annotations`, the renderer knows nothing about a game's notation. You must provide instead the column and row numbers, which are zero-based: 0,0 is the top row, top column.
          *
          * @minItems 1